changed all PR comments
fixed enable/disable/stop (also added enable to the end() of MoveLift so
it's ready and live when not using that non-PID command anymore
changed getString in OI to be SmartDashboard. instead of Robot. but left
my getString methods in there bc not doing any harm
left end() in interrupted bc not really doing any harm
got rid of spd --> voltage method and replaced it with a constant from
Robot.getConst